Fixing merging issues that came up
